refactor(services): separada a camada de services
A camda de services, que anteriormente estava acoplada com as rotas, foi separada

fix: #9
