+++ src/routes.js
@@ -1,19 +1,12 @@
-const Koa = require('koa')
 const koaBody = require('koa-body')
 const koaRouter = require('koa-router')
-const createCardDB = require('./db')
-
+const services = require('./services')
 const router = new koaRouter()
 
-const createCardRequest = (ctx) => {
-  const receivedRequestBody = ctx.request.body
-  createCardDB.createNewCard(receivedRequestBody)
-  ctx.body = ctx.request.body
-}
-
-router.post('/card', koaBody(), createCardRequest)
-router.get('/card', (ctx) => {
-  ctx.body = 'Hey amigo'
+router.post('/card', koaBody(), (ctx) => {
+  const requestBody = ctx.request.body
+  ctx.body = requestBody
+  services.createCardRequest(requestBody)
 })
 
 module.exports = router
